[FIX] nav btn not popping on scroll in bigger devices
